Handle delete events
Handle delete events for ClusterIngress objects.

Set OwnerReference to router namespace

In order to trigger a cascade deletion for the kubernetes garbage
collector the router namespace needs the OwnerReference field set to the
Cluster Ingress Operator object.

Change deletion pattern

The previous deletion pattern relayed on a non supported
garbage collection functionality, this patch changes the
pattern by setting a label on the serivce and router
deployment and search for this label when the delete
event is triggered.

Remove OwnerRef code

Delete ClusterIngress objects from manifests

Set Onwer Reference to the service

By setting the DaemonSet as ownerReference to the service
we let the k8s garbage collector delete it when deleting
the ClusterIngress object.

Fix logging and error handling

Code cleanup

Code cleanup