Publish operator status
Create (if necessary) and update a ClusterOperator object to indicate the
operator's status.

* cmd/cluster-ingress-operator/main.go: Add a watch on daemonsets in the
application namespace to trigger reconciliation on status updates.
* manifests/00-cluster-role.yaml: Allow the operator to create, get, and
update clusteroperators.
* pkg/manifests/bindata.go: Regenerate.
* pkg/util/clusteroperator/status.go: Add new functions: SetStatusCondition
and ConditionsEqual.
* pkg/util/clusteroperator/status_test.go: Add tests.
* pkg/stub/status.go: Add new functions: syncOperatorStatus,
getOperatorState, and computeStatusConditions.
Use getOperatorState in syncOperatorStatus to get the prerequisite
state for computing the operator's status.
Use computeStatusConditions in syncOperatorStatus to compute the status.
Use clusteroperator.ConditionsEqual in syncOperatorStatus to determine
whether an update is needed.
Use clusteroperator.SetStatusCondition in computeStatusConditions to
compute new conditions.
* pkg/stub/handler.go: Use syncOperatorStatus in Handle to update
ClusterOperator.
* pkg/stub/handler_test.go: Add tests.