Add UserDefinedClusterIngressHA
Add a high-availability strategy that exposes ports 80 and 443 as endpoints
on the host network and leaves implementation of high-availability up to
the user.

* pkg/apis/ingress/v1alpha1/clusteringress_types.go: Define
UserDefinedClusterIngressHA.
* pkg/manifests/manifests.go (RouterDeployment): If HA type is
UserDefinedClusterIngressHA, use host network, and configure probes to use
localhost (defaulting to the node IP address could cause problems if it is
not routable or is blocked by a firewall).
* pkg/manifests/manifests_test.go: Test with nil, CloudClusterIngressHA,
and UserDefinedClusterIngressHA.

This commit resolves NE-116.

https://jira.coreos.com/browse/NE-116