Make clusteringress scalable
Add /scale and /status subresources to clusteringress, add fields to spec
and status to facilitate scaling, and modify the controller to scale the
router in accordance with the clusteringress.

This commit implements NE-127.

https://jira.coreos.com/browse/NE-127

* assets/defaults/cluster-ingress.yaml: Set default replicas value to 1.
* manifests/00-cluster-role.yaml: Allow operator to update clusteringress
status.
* manifests/00-custom-resource-definition.yaml: Add /scale and /status
subresources.
* pkg/apis/ingress/v1alpha1/clusteringress_types.go (ClusterIngressSpec):
Add replicas field.
(ClusterIngressStatus): Add replicas and selector fields.
* pkg/manifests/bindata.go: Regenerate.
* pkg/manifests/manifests.go (RouterDeployment): Copy replicas value from
the clusteringress spec to the router deployment spec.
* pkg/manifests/manifests_test.go (TestManifests): Make sure deployment's
replicas value gets set to the value from the clusteringress or to an
appropriate default.
* pkg/operator/controller/controller.go (ensureRouterForIngress): Update
the clusteringress's status, using syncClusterIngressStatus.
(syncClusterIngressStatus): New function that updates the given
clusteringress's status with the current replicas and selector.
(deploymentConfigChanged): Check the deployment's replicas field.
* test/e2e/operator_test.go (TestClusterIngressScale): New test that checks
the clusteringress's status and tries to scale the router up and down using
the scale client.
(getScaleClient): New helper function for TestClusterIngressScale.