Use self-signed default router certificate
Generate a CA certificate and use it to generate default certificates for
routers.  Publish the CA certificate in a configmap, which can be
incorporated into CA bundles in order to trust the default certificates.

This commit resolves NE-139.

https://jira.coreos.com/browse/NE-139

* manifests/00-cluster-role.yaml: Grant the operator create, get, list,
watch, and delete access to secrets.
* pkg/manifests/manifests.go (RouterServiceInternal): Change serving-cert
secret name so as not to conflict with the secret that the operator
creates.
* pkg/manifests/manifests_test.go (TestManifests): Adjust for the change to
RouterServiceInternal.
* pkg/operator/controller/controller.go (caCertSecretName):
(caCertConfigMapName): New constants for the name of the secret and
configmap for the CA certificate.
(CACert): New type to hold a cached copy of the CA certificate and to keep
track of the resource versions of the most recently successfully observed
secret and configmap for the CA certificate.
(Config): Add CACert field.
(ensureRouterForIngress): Call ensureDefaultCertificateForIngress.
(ensureDefaultCertificateForIngress): New function to check that a default
certificate exists for a given ClusterIngress, and generate a certificate
and create a secret for it if none exists.  Use ensureRouterCACertificate
to get the certificate used to sign the default certificate.
(ensureRouterCACertificate): New function to get the CA certificate for
signing default certificates, and generate a CA certificate and create a
configmap and secret for it if none exists.
* test/e2e/operator_test.go (TestRouterCACertificate): New test that gets
the configmap with the CA certificate and verifies that the router's
default certificate is valid using the CA certificate.