Delete router-ca configmap if it is not needed
If there exists no clusteringress that uses the generated default
certificate, delete the configmap for the CA certificate.

This change makes it straightforward for other components to incorporate
cluster-ingress-operator's CA certificate into their trust bundles exactly
when some router has a default certificate signed by the CA certificate:
The router-ca configmap will exist in the openshift-config-managed
namespace iff the certificate CA should be trusted.

This is a follow-up to commit 228c9b30dc89f7186e481b30c2f22e781073fd4d.

Related to NE-139.

https://jira.coreos.com/browse/NE-139

* pkg/operator/controller/controller.go (reconcile): Use the new
shouldPublishRouterCA function to check whether the CA certificate should
be published in a configmap, and use the new ensureRouterCAIsPublished
method create or update it and the new ensureRouterCAIsUnpublished method
to delete it as appropriate.
(ensureRouterForIngress): Use the new ensureDefaultCertificateDeleted
method to delete the operator-generated default certificate if
.spec.defaultCertificateSecret is set on the clusteringress.
(ensureDefaultCertificateForIngress): Delete unnecessary comment.  Replace
ensureRouterCACertificate with getRouterCA.
(ensureDefaultCertificateDeleted): New function that deletes the
operator-generated default certificate.
(getRouterCA): New function to get the CA, or create it if it does not
already exist, using ensureRouterCACertificate.
(ensureRouterCACertificate): Rename from this...
(ensureRouterCACertificateSecret): ...to this.  Change from returning the
CA itself to returning the secret.  Delete unnecessary comment.  Do not
create the configmap, which is now the responsibility of the reconcile
function.
(shouldPublishRouterCA): New function that returns a Boolean value
indicating whether there exists any clusteringress that uses the generated
default certificate (in which case the CA certificate that was used to
generate them should be published).
(ensureRouterCAIsPublished): New function that creates or updates the
configmap as needed.
(ensureRouterCAIsUnpublished): New function that deletes the configmap.
* pkg/operator/controller/controller_test.go (TestShouldPublishRouterCA):
New test for shouldPublishRouterCA.
* test/e2e/operator_test.go (TestClusterIngressUpdate): Make sure that the
configmap exists for the default clusteringress, is deleted when the
clusteringress has a custom default certificate set, and is recreated when
the clusteringress is changed back to using a default certificate generated
by the operator.