Set requests on operator and router deployments
Set resources.requests.cpu=10m on the operator's container, and set
resources.requests.cpu=100m and resources.requests.memory=256Mi on the
router's container.

The setting for the operator is the standard setting for operators that run
on masters.  The setting for the router is the default in Origin 3.11 (see
<https://github.com/openshift/origin/blob/b054a7bf5b11c760e5dd0f870fc852cc9b8b748b/pkg/oc/cli/admin/router/router.go#L907-L912>).

This commit resolves NE-142.

https://jira.coreos.com/browse/NE-142

* assets/router/deployment.yaml: Set resources.requests.cpu=100m and
resources.requests.memory=256Mi on the router container.
* manifests/02-deployment.yaml: Set resources.requests.cpu=10m on the
operator container.
* pkg/manifests/bindata.go: Regenerate.