dns/aws: update public zone discovery to search parent domains
The installer needs to move towards [1] where the private zone named as `cluster_name.base_domain` ie. `cluster_domain` because of [2], but the
public zone still remains `base_domain` as that cannot be created by instaler.

This means that cluster-ingress-operator cannot use the public r53 zone with the same name as the `base_domain` from `DNS.config.openshift.io` [3] as it will be set to the
`cluster_domain`.

This changes the public zone discovery to find a public zone which is the nearest parent domain to `cluster_domain`.

[1]: https://github.com/openshift/installer/pull/1169
[2]: https://github.com/openshift/installer/issues/1136
[3]: https://github.com/openshift/api/blob/d67473e7f1907b74d1f27706260eecf0bc9f2a52/config/v1/types_dns.go#L28