controller: clean up error handling
When reconciling a clusteringress, establish the invariant that DNS and
Infrastructure cluster config must exist; this makes it easier to deal with
error handling and there's little we can do without cluster config anyway. Make
it possible to do some additional work when not able to reconcile a
clusteringress. Pass DNS and Infrastructure around where needed for consistency
with the other refactored bits.