Move router customization out of manifest factory
Move the logic that customizes the router deployment out of the manifest
factory and into the controller.

* pkg/manifests/manifests.go (RouterDeployment): Rename...
(RouterDeploymentAsset): ...to this.  This change is necessary in order to
be able to use the name for a function (see next change).
(RouterDeployment): Drop the receiver (making this a function) and the
error return value.  Panic if unable to read the asset.  Otherwise return
the unmodified asset.  Move all customization of the deployment from
here...
* pkg/operator/controller/controller_router_deployment.go
(desiredRouterDeployment): ...to here.  New function that returns the
desired router deployment based on the provided clusteringress, router
image name, and infrastructure config.
(routerDeploymentName): New function to generate the namespaced name for
the router deployment.
(ensureRouterDeployment): Add parameter for infrastructure config.  Use
desiredRouterDeployment to get the expected deployment.
(ensureRouterDeleted): Adjust to the factory function's new signature.  Use
routerDeploymentName to set the deployment's name and namespace.
* pkg/operator/controller/controller.go (ensureClusterIngress): Pass
infraConfig to ensureRouterDeployment.
* pkg/manifests/manifests_test.go (TestManifests): Move checks on the
router deployment from here...
* pkg/operator/controller/controller_router_deployment_test.go
(TestDesiredRouterDeployment): ...to here.