Fix high availability defaulting and status
Move the high availability configuration defaulting from the manifest
factory to the controller.  Reflect the actual HA configuration in the
status of the clusteringress rather than setting it in the spec.

This commit resolves NE-157.

https://jira.coreos.com/browse/NE-157

* pkg/apis/ingress/v1alpha1/clusteringress_types.go (NoClusterIngressHA):
New constant.  This HA type corresponds to the behavior when
.spec.highAvailability was nil prior to this commit.
(ClusterIngressStatus): Add HighAvailability field.
* pkg/apis/ingress/v1alpha1/zz_generated.deepcopy.go: Regenerate.
* pkg/manifests/manifests.go (DefaultClusterIngress): Drop the receiver
and the error return value.  Panic if unable to read the asset.  Otherwise
return the unmodified asset.
* pkg/manifests/manifests_test.go (TestDefaultClusterIngress): Delete
checks for .spec.highAvailability.
* pkg/operator/controller/controller.go (reconcile): Pass infraConfig to
ensureIngressDeleted.
(haTypeForInfra): New function to determine the HA type from the
infrastructure config.
(effectiveHighAvailability): New function to determine the HA configuration
from a clusteringress and infrastructure config.  Determine the current HA
type from the clusteringress spec if it is specified there or else using
haTypeForInfra.
(ensureIngressDeleted): Add parameter for infrastructure config.  Use
expectedHighAvailability to determine the HA configuration.  Pass the HA
configuration to finalizeLoadBalancerService.
(ensureClusterIngress): Use effectiveHighAvailability to determine the
desired HA configuration.  Pass it to ensureRouterDeployment,
ensureLoadBalancerService, ensureDNS, and syncClusterIngressStatus.
(syncClusterIngressStatus): Add parameter for HA.  Set the current HA
status in the clusteringress status.
* pkg/operator/controller/controller_dns.go (ensureDNS): Add parameter for
HA.  Use this parameter to pass the HA type to desiredDNSRecords.
(desiredDNSRecords): Add parameter for HA type.  Use this parameter instead
of using the clusteringress spec to determine whether to manage DNS.
* pkg/operator/controller/controller_lb.go (ensureLoadBalancerService): Add
parameter for HA.  Use this parameter to pass the HA type to
desiredLoadBalancerService.
(desiredLoadBalancerService): Add parameter for HA type.  Use this
parameter instead of using the clusteringress spec to determine whether to
create an LB service.
(finalizeLoadBalancerService): Add parameter for HA.  Use this parameter to
pass the HA type to desiredDNSRecords.
* pkg/operator/controller/controller_router_deployment.go
(ensureRouterDeployment): Add parameter for HA.  Use this parameter to pass
the HA type to desiredRouterDeployment.
(desiredRouterDeployment): Add parameter for HA type.  Use this parameter
instead of using the clusteringress spec to determine whether to use PROXY
protocol or host networking.
* pkg/operator/controller/controller_router_deployment_test.go
(TestDesiredRouterDeployment): Pass appropriate test inputs for the HA type
to desiredRouterDeployment.
* pkg/operator/operator.go (Start): Adjust to the new signature of
DefaultClusterIngress.
* test/e2e/operator_test.go (TestRouterCACertificate): Get HA type from
clusteringress status.