Use object from API when recording an event
Avoid "Could not construct reference to: [...] due to: 'selfLink was empty,
can't make reference'. Will not report event: [...]" errors from the event
recorder by using an object we have gotten from the API, as opposed to one
we have constructed ourselves (which is missing the self link).

* pkg/operator/controller/certificate/ca.go (ensureRouterCASecret): After
creating the secret, get the newly created object from the API.  Use the
object from the API for the event recorder and for the return value.
* pkg/operator/controller/certificate/publish_ca.go
(ensureRouterCAConfigMap): After creating the configmap, get the newly
created object from the API, and use this object for the event recorder.
When updating the configmap, use the object we got earlier from the API for
the event recorder.
* pkg/operator/controller/certificate/publish_certs.go
(ensureRouterCertsGlobalSecret): After creating the secret, get the newly
created object from the API, and use this object for the event recorder.
When updating the secret, use the object we got earlier from the API for
the event recorder.