api: refactor to ingresscontroller.operator.openshift.io API
Replace the v1alpha1 clusteringress.ingress.openshift.io type with the new
ingresscontroller.operator.openshift.io API from openshift/api.

This commit is intended to be the absolute minimum work to switch to the new
API. Internal references to clusteringresses, routers, and any user-facing
references to things renamed in the new API are not covered by this commit.