Add certificate-publisher-controller
Extract a new "certificate-publisher-controller" out of the existing
certificate-controller.  The new controller watches ingresscontrollers in
the operator namespace and default-certificate secrets in the operand
namespace and updates the router-certs secret in the
openshift-config-managed namespace.

Because certificate-controller watches ingresscontrollers and not secrets,
it was failing to update router-certs in some situations where the secret
needed to be updated.  In particular, creating, updating, or deleting a
secret _after_ creating an ingresscontroller that references the secret
should cause router-certs to be updated.  The new controller ensures that
router-certs is updated in this scenario.

This commit fixes bug 1687931.

https://bugzilla.redhat.com/show_bug.cgi?id=1687931

* pkg/operator/controller/certificate/controller.go (Reconcile): Remove
code to publish the router-certs secret.
* pkg/operator/controller/certificate-publisher/controller.go: New file.
Define a new controller that manages the router-certs secret.
* pkg/operator/controller/certificate/publish_certs.go:
* pkg/operator/controller/certificate/publish_certs_test.go: Move to
pkg/operator/controller/certificate-publisher/.
* pkg/operator/operator.go (New): Initialize the new controller.
* test/e2e/certificate_publisher_test.go: New file.
(TestCreateIngressControllerThenSecret):
(TestCreateSecretThenIngressController): New tests.
* test/e2e/operator_test.go (createDefaultCertTestSecret): Add a parameter
for the name.  Swap the return values to make the error the second value.
(TestClusterIngressUpdate): Pass createDefaultCertTestSecret a generated
name.