Fix load-balancer finalization
When finalizing a load-balancer service, check whether it has been
provisioned, and if not, skip cleaning up its DNS records.

Before this commit, if an ingresscontroller was created with the
LoadBalancerService endpoint publishing strategy and subsequently deleted
before the load balancer had been provisioned, the finalization logic
returned an error, which caused the controller to retry and fail to
finalize the service repeatedly.

* pkg/operator/controller/controller_dns.go (ensureDNS): Check whether the
load balancer has been provisioned.  Return an error if not.  Pass the
host name from the load balancer to desiredDNSRecords.
(desiredDNSRecords): Take a host name instead of a load-balancer service.
* pkg/operator/controller/controller_lb.go (finalizeLoadBalancerService):
Check whether the load balancer has been provisioned.  Skip DNS clean-up if
not.  Pass the host name from the load balancer to desiredDNSRecords.