deploymentConfigChanged: Fix deleting tolerations
Always update tolerations if the router deployment has changed,

Before this commit, deploymentConfigChanged only updated the tolerations on
the deployment if the deployment was expected to have tolerations.  As a
result, deploymentConfigChanged did not remove tolerations when the
deployment had tolerations and was expected not to.

This commit fixes bug 1701462.

https://bugzilla.redhat.com/show_bug.cgi?id=1701462

* pkg/operator/controller/controller_router_deployment.go
(deploymentConfigChanged): Delete condition for assignment of tolerations.