Choose an appropriate default deployment strategy for an endpoint publishing strategy
Choose an appropriate default deployment strategy for an endpoint publishing
strategy.

The default is a rolling update with 25% max unavailable and 25% max surge. This
enables 100% availability rollouts at any replica/node count when the publishing
strategy supports colocated ingress controllers.

With HostNetwork, use 25% max unavailable and 0 surge. This is because the
ingress controllers can't be colocated. Usually, replicas will be equal to the
node pool size. Under these conditions, surge requires new nodes to support the
rollout. This means a positive surge can cause the rollout to wedge in the
absence of auto-scaling.