Make deployment topologies more predictable by prohibiting co-location
Before this change, ingresscontrollers published via LoadBalancerService could
be co-located according to a soft anti-affinity policy. This policy resulted in
a high probability of scheduler outcomes whereby 100% of ingresscontroller
processes were co-located on a single node after cluster installation, defeating
HA.

In the absence of the descheduler (for auto-rebalancing), the possibility
of unwanted co-location is too great a risk given admins have no obvious way
to detect the surprise co-location and no good procedure for forcing rebalancing.

After this change, deployment topologies are made more predictable by disabling
co-location for all publishing strategies.

An additional consequence of this change is that 100% availability rollouts when
the number of deployment replicas is equal to the number of nodes are no longer
possible even when using LoadBalancerService. See fb9851a for more details on
the rationale.