TestIngressControllerUpdate: Re-get between updates
Get the ingress controller from the API in between updates so that that
each update uses the latest resource version in order to prevent the
following test failure:

    --- FAIL: TestIngressControllerUpdate (12.77s)
    	operator_test.go:367: failed to reset IngressController: Operation cannot be fulfilled on ingresscontrollers.operator.openshift.io "default": the object has been modified; please apply your changes to the latest version and try again

This commit fixes bug 1705100.

https://bugzilla.redhat.com/show_bug.cgi?id=1705100

* test/e2e/operator_test.go (TestIngressControllerUpdate): Wait for ingress
controller to be available at the start of the test.  Get the ingress
controller in between updates to ensure we have the latest resource version
for each update.  Use RouterOperatorGeneratedDefaultCertificateSecretName
instead of hard-coding the default secret name.  Minor code clean-ups.