Surface service-controller LB provisioning failures through status
When an LB is determined to be pending, try and surface more detail by analyzing
events in the operand namespace related to the LB service, and if an LB creation
failure event is detected, propagate the message into IngressController status
and provide a more specific reason.

Replace the use of indexers with simpler inline cache lookups now that the
manager cache is available.