Honor the load balancer strategy's scope parameter
Respect the "scope" parameter for the ingress controller's load balancer
strategy in order to allow choosing between an external (the default) or an
internal load balancer.

This commit resolves NE-167.

https://jira.coreos.com/browse/NE-167

* pkg/operator/controller/controller.go
(enforceEffectiveEndpointPublishingStrategy): Default
status.endpointPublishingStrategy.loadBalancer.scope to "External".
* pkg/operator/controller/controller_lb.go (internalLBAnnotations): New
variable.  Map platform type to the annotation for that platform that makes
the load balancer internal-only.
(desiredLoadBalancerService): Check
status.endpointPublishingStrategy.loadBalancer.scope and use
internalLBAnnotations to set the appropriate annotation if the scope is not
"External".
* pkg/operator/controller/dns/controller.go (Reconcile): Omit the public
zone if the load balancer is internal-only.
* test/e2e/operator_test.go (TestInternalLoadBalancer): New test.