getPlatformStatus: Fix platform upgrading from 4.1
Use the deprecated status.platform field from the cluster infrastructure
config to determine the platform if the status.platformStatus field is nil.
In particular, platformStatus will be nil on a cluster that is being
upgraded from 4.1.

This commit fixes bug 1734414.

https://bugzilla.redhat.com/show_bug.cgi?id=1734414

* cmd/ingress-operator/main.go (getPlatformStatus): Use the status.platform
field from the cluster infrastructure config as the platform type if
status.platformStatus is nil.