status: compute ingresscontroller availability from deployment conditions
Before this change, ingresscontroller availability was determined based on
deployment available replicas from status. Now, consider the ingresscontroller
available if the deployment has an `Available=True` condition. Available
replicas is a weak proxy for deployment availability. The condition takes into
account the complex availability rules of a deployment and also provides
additional context as to why the condition is in a certain state.